First part of refactoring for graph gen tool
